requirements.txt
Pillow>=10.1.0
requests
numpy
torch # ComfyUI lo requiere